Fix namespace collision with helper module
Once the custom resource is defined, there exists a class of
Chef::Resource::ChefHandler, which conflicts with the
ChefHandler::Helpers module namespace. This typically doesn't matter
except in contexts like ChefSpec, where the resource might be redefined
over the course of the tests. In this case, the inclusion of the helper
module conflicts with the namespace of the already-defined resource.

This is fixed by explicitly including the toplevel ::ChefHandler
namespace for the helper module.

Resolves #62

Signed-off-by: Chris Crebolder <ccrebolder@gmail.com>